Adds auto indent strategy for multi line strings
To make this component possible, it was necessary to disable the
"indent on tab" action of the Java editor, which is done in the
`ScalaSourceFileEditor`.

The Java editor, it its wisdom, handled the tab key entirely by itself
and made it impossible to execute further components when this key is
pressed. Because this was special and even more weird behavior in
comparison on how other key events are handled by the Scala editor it
seemed to be the best to disable the behavior. The downside of this is
that now an additional component is required that handles the tab key.

There is another action, simply called "Indent" by the Java editor,
that handles indentation when a newline key is inserted and the
following line has to be indented. Because this behavior is buggy for
Scala code in a lot of cases it is necessary to disable this action as
well and implement a more Scala aware indentation logic. But this work
is not part of this commit.

The new features can be disabled in the editor preference page. See the
containing tests for a specification of the behavior, which not only
allows to do auto indentation inside of multi line strings but also the
automatic insertion of a stripMargin call to ease correct indentation
of multi line strings.

Fixes #1001099, #1001670, #1002063