Ease condition to show only tooltips in code completion
On constructor completion only the tooltip, which shows the name of all
the parameters that need to be inserted, should be shown. This may not
be the best behavior but this is how it works at the moment. The
condition that checks this can safely abort the whole method that should
insert the actual completion because there is nothing to insert.