Replace references to Java partitions with references to Scala partitions
This shouldn't introduce any functional changes beside from the point
that everything related to partitioning (syntax highlighting, auto
edits) have a chance to not work anymore.

Some new Scala partitions are added, they replace partitions located in
`IJavaPartitions` and `IDocument`.

Because this is a no compile-time-save change, it could be that some
references are forgotten, but hopefully my grep-foo caught them all.