Convert constants of UPPER_CASE to CamelCase style
CamelCase style conforms to the official Scala style guide and helps to
distinguish better between Scala and Java API in some few cases.

All changes are mostly done automatically with help of find, grep, awk
and vim. The naming of some identifiers was adjusted in some cases in
order to read them more easily.

Nearly all changes belong to the `core` and `core.test` projects - the
debugger was not touched because I'm not familiar with this code.

Beside from breaking all plugins that depend on the Scala IDE use any
any of its constants, this shouldn't introduce semantic changes to the
codebase (thanks to scalac for veryving that this holds).