Start implementing API for Scala IDE extensions
This introduces several new traits in the new package

  org.scalaide.core.text

The main design goal of the API is to provide a safe environment users
can work on. The entire API can't change the underlying documents, it
can only produce change objects that need to be passed to the IDE, which
can decide what to do with the changes.